requirements.txt
# Core app and bundled tool dependencies.
# This file intentionally includes the Python packages needed by:
# - the SelectInfo app itself
# - the MCP server
# - the bundled OneForAll and dirsearch tool sources
#
# External binaries such as nmap and subfinder are not pip packages and
# still need to be installed or configured separately.

beautifulsoup4>=4.12.0
cdx_toolkit>=0.9.38
certifi>=2022.6.15
charset-normalizer>=2.0.0
colorama>=0.4.4
defusedxml>=0.7.0
dnspython>=2.2.1
exrex>=0.11.0
fire>=0.5.0
future>=0.18.2
Jinja2>=3.0.0
loguru>=0.6.0
mcp>=1.23.3,<2
mysql-connector-python>=8.0.20
ntlm-auth>=1.5.0
psycopg[binary]>=3.0
pyOpenSSL>=21.0.0
pyparsing>=2.4.7
PySocks>=1.7.1
requests>=2.31.0
requests-ntlm>=1.1.0
requests-toolbelt>=1.0.0
rich>=13.7.0
setuptools
SQLAlchemy>=1.3.22
tenacity>=8.0.1
tldextract>=3.1.2
tqdm>=4.60.0
treelib>=1.6.1
